Downloads/pdf-silin-worker-main/requirements/base.txt
# Base
pytz==2020.5
python-slugify==4.0.1
Pillow==8.2.0
boto3==1.17.96
redis==4.1.2
unidecode==1.3.8

# Django
Django==3.1.5

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.12.2
django-filter==2.4.0
django-cors-headers==3.7.0

# Tools
django-extensions==3.1.0

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==5.2.0

# DOCS
docxtpl==0.11.5
python-barcode==0.13.1
treepoem==3.17.0
jsonpath-ng==1.5.3

# PDF
unoserver==1.1

# HTTP Server
gunicorn==20.0.4

# requests
requests==2.29.0
backoff==2.2.1

# JWT
cffi==1.15.1
pycparser==2.21
PyJWT==2.8.0
cryptography==41.0.2